Use bower for libs. Track caman.js at 4.1.2
gitignore the lib folder but commit required lib files with `git add -f`. This
is a simple way to have a full bower folder in lib/ but ignore all the files
that aren't needed. All the installed files will still be included in the _site
but at least the commit history will be simple.

Don't copy bower.json or Makefile into the _site folder: they're not needed.
Perhaps more could be excluded.